requirements.txt
slack-bolt==1.27.0
aiohttp>=3.9,<4.0
cachetools>=5.0,<6.0
python-dotenv>=1.0,<2.0
structlog>=24.0,<25.0
claude-agent-sdk==0.1.49
asyncpg>=0.29,<1.0
aiosqlite>=0.21.0
anthropic>=0.49
google-cloud-secret-manager>=2.18,<3.0
